# front/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AdoptMe API Documentation",
    "description": "Especificación interactiva de los endpoints de la plataforma de adopción de mascotas AdoptMe. Incluye gestión de usuarios, roles, catálogo de mascotas y solicitudes de adopción.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo local"
    }
  ],
  "paths": {
...
    "/api/mascotas": {
      "get": {
        "summary": "Listar mascotas",
        "description": "Retorna la lista de mascotas registradas. Puede filtrarse para obtener solo las de una empresa específica.",
        "parameters": [
          {
            "name": "mine",
            "in": "query",
            "description": "Flag para filtrar las mascotas de la empresa autenticada ('1')",
            "required": false,
            "schema": { "type": "string", "example": "1" }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "ID de la empresa para filtrar si mine es '1'",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "rol",
            "in": "query",
            "description": "Rol del usuario que realiza la consulta ('empresa')",
            "required": false,
            "schema": { "type": "string", "example": "empresa" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mascotas obtenida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mascota"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Publicar una mascota",
        "description": "Registra una nueva mascota disponible para adopción (Rol Empresa).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string", "example": "Max" },
                  "especie": { "type": "string", "enum": ["perro", "gato"], "example": "perro" },
                  "raza": { "type": "string", "example": "Pastor Alemán" },
                  "edad": { "type": "integer", "example": 2 },
                  "descripcion": { "type": "string", "example": "Muy juguetón y cariñoso." },
                  "estado": { "type": "string", "enum": ["disponible", "adoptada"], "default": "disponible" }
                },
                "required": ["nombre", "especie"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mascota creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mascota"
                }
              }
            }
          },
          "400": {
            "description": "Nombre obligatorio o especie no válida"
          }
        }
      }
    },
    "/api/mascotas/{id}": {
      "get": {
        "summary": "Obtener mascota por ID",
        "description": "Retorna los detalles de una mascota específica.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único de la mascota",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de mascota obtenidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mascota"
                }
              }
            }
          },
          "404": {
            "description": "Mascota no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar mascota",
        "description": "Modifica la información de una mascota (Rol Empresa).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la mascota a actualizar",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string", "example": "Max Actualizado" },
                  "edad": { "type": "integer", "example": 3 },
                  "especie": { "type": "string", "example": "perro" },
                  "raza": { "type": "string", "example": "Pastor Alemán" },
                  "descripcion": { "type": "string", "example": "Sigue buscando hogar." },
                  "estado": { "type": "string", "example": "disponible" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mascota actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mascota"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud vacía o campos incorrectos"
          },
          "404": {
            "description": "Mascota no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar mascota",
        "description": "Elimina a una mascota del catálogo.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de la mascota",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Mascota eliminada correctamente"
          },
          "404": {
            "description": "Mascota no encontrada"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Mascota": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nombre": { "type": "string", "example": "Luna" },
          "especie": { "type": "string", "example": "gato" },
          "raza": { "type": "string", "example": "Siames" },
          "edad": { "type": "integer", "example": 1 },
          "descripcion": { "type": "string", "example": "Muy mimosa." },
          "estado": { "type": "string", "example": "disponible" }
        }
      },
...
    }
  }
}